mcp_servers/python/servers/MCP-STOCKANALYZER/mcp-stockanalyzer/pyproject.toml
[project]
name = "mcp-stockanalyzer"
version = "0.1.0"
description = "MCP server for stock analysis and news."
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
 "yfinance",
 "newsapi-python",
 "numpy",
 "mcp",
 "fastapi",
]
[[project.authors]]
name = "Your Name"
email = "your@email.com"

[build-system]
requires = [ "hatchling",]
build-backend = "hatchling.build"

[dependency-groups]
dev = [
    "pyright",
]

[project.scripts]
mcp-stockanalyzer = "mcp_stockanalyzer:main" 